requirements.txt
aiofiles==23.2.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.2.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cross-web==0.4.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
f==0.0.1
fastapi==0.109.0
graphql-core==3.2.7
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.3
kombu==5.6.2
lxml==6.0.2
Markdown==3.10.1
MarkupSafe==3.0.3
neo4j==5.16.0
numpy==1.26.3
ollama==0.1.6
openpyxl==3.1.5
packaging==25.0
pandas==3.0.0
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.52
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
PyPDF2==3.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.35.1
strawberry-graphql==0.289.7
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.27.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.5.0
websockets==15.0.1
xlrd==2.0.2
psutil==5.9.8
